Skiller feilhåndtering ut til egen fil
